Massively improve file structure
