+++ src/CSF.Core/Abstractions/Attributes/PreconditionAttribute.cs
