+++ src/CSF.Core/Abstractions/Attributes/Preconditions/IPrecondition.cs
