+++ src/CSF.Core/Abstractions/Components/IComponent.cs
