+++ src/CSF.Core/Abstractions/Components/IConditionalComponent.cs
