+++ src/CSF.Core/Abstractions/Context/IConsoleContext.cs
