+++ src/CSF.Core/Abstractions/Context/IContext.cs
