+++ src/CSF.Core/Abstractions/Parsing/ParseInformation.cs
