+++ src/CSF.Core/Abstractions/Results/IResult.cs
@@ -23,12 +23,4 @@ namespace CSF
         /// </summary>
         public Exception Exception { get; }
     }
-
-    public static class ResultExtensions
-    {
-        public static ValueTask<T> AsValueTask<T>(this T value)
-        {
-            return new ValueTask<T>(value);
-        }
-    }
 }
