+++ src/CSF.Core/Abstractions/TypeReaders/ITypeReader.cs
