+++ src/CSF.Core/CommandManager.cs
@@ -275,82 +275,4 @@ namespace CSF
         protected virtual CheckResult OnBestOverloadUnavailable(IContext context)
             => CheckResult.FromError($"Failed to find overload that best matches input: {context.Name}.");
     }
-
-    /// <summary>
-    ///     
-    /// </summary>
-    public static class CommandManagerExtensions
-    {
-        /// <summary>
-        ///     
-        /// </summary>
-        /// <param name="collection"></param>
-        /// <param name="action"></param>
-        /// <returns></returns>
-        public static IServiceCollection AddCommandManager(this IServiceCollection collection, Action<ManagerBuilderContext> action = null)
-        {
-            collection.AddCommandManager<CommandManager>(action);
-
-            return collection;
-        }
-
-        /// <summary>
-        ///     
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="collection"></param>
-        /// <param name="action"></param>
-        /// <returns></returns>
-        public static IServiceCollection AddCommandManager<T>(this IServiceCollection collection, Action<ManagerBuilderContext> action = null)
-            where T : CommandManager
-        {
-            var context = new ManagerBuilderContext();
-
-            action?.Invoke(context);
-
-            collection.AddCommandManager<T>(context);
-
-            return collection;
-        }
-
-        /// <summary>
-        ///     
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="collection"></param>
-        /// <param name="context"></param>
-        /// <returns></returns>
-        public static IServiceCollection AddCommandManager<T>(this IServiceCollection collection, ManagerBuilderContext context)
-            where T : CommandManager
-        {
-            collection.AddComponents(context);
-            collection.AddTypeReaders(context);
-
-            collection.TryAddSingleton<T>();
-
-            return collection;
-        }
-
-        public static IEnumerable<T> CastWhere<T>(this IEnumerable input)
-        {
-            foreach (var @in in input)
-                if (@in is T @out)
-                    yield return @out;
-        }
-
-        public static IReadOnlyList<T> GetRange<T>(this IReadOnlyList<T> input, int startIndex, int? count = null)
-        {
-            IEnumerable<T> InnerGetRange()
-            {
-                count ??= (input.Count - startIndex);
-
-                for (int i = startIndex; i <= count; i++)
-                    yield return input[i];
-            }
-
-            var range = InnerGetRange();
-
-            return range.ToList();
-        }
-    }
 }
