+++ src/CSF.Core/Helpers/ManagerBuildingContext.cs
