+++ src/CSF.Core/Implementations/Attributes/CommandAttribute.cs
