+++ src/CSF.Core/Implementations/Attributes/ComplexAttribute.cs
