+++ src/CSF.Core/Implementations/Attributes/DontRegisterAttribute.cs
