+++ src/CSF.Core/Implementations/Attributes/GroupAttribute.cs
