+++ src/CSF.Core/Implementations/Attributes/Preconditions/RequireContextAttribute.cs
