+++ src/CSF.Core/Implementations/Components/Command.cs
