+++ src/CSF.Core/Implementations/Components/Helpers/ComponentContainer.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace CSF
+{
+    public sealed class ComponentContainer
+    {
+        public IEnumerable<IConditionalComponent> Values { get; }
+
+        public ComponentContainer(IEnumerable<Type> types)
+            => Values = types.SelectMany(x => new Module(x).Components);
+    }
+}
