+++ src/CSF.Core/Implementations/Components/Module.cs
