+++ src/CSF.Core/Implementations/Parsing/TextParser.cs
