+++ src/CSF.Core/Implementations/Results/ArgsResult.cs
