+++ src/CSF.Core/Implementations/Results/CheckResult.cs
