+++ src/CSF.Core/Implementations/Results/CheckYieldResult.cs
