+++ src/CSF.Core/Implementations/Results/ExecuteResult.cs
