+++ src/CSF.Core/Implementations/Results/PreconditionResult.cs
