+++ src/CSF.Core/Implementations/Results/SearchResult.cs
