+++ src/CSF.Core/Implementations/Results/TypeReaderResult.cs
