+++ src/CSF.Core/Implementations/TypeReaders/BaseTypeReader.cs
