+++ src/CSF.Core/Implementations/TypeReaders/ColorTypeReader.cs
