+++ src/CSF.Core/Implementations/TypeReaders/EnumTypeReader.cs
